fix for issue #128
Fatal error when  trying to delete empty category from repository...
because of call to non existing Function "contain_subcategories_objects($category_id)"

The function allowed_to_delete_category() can be reduced to return true;
as it has no impact to the actual delete process at all.

The impactViewer is taking care of that.